Update getContactsByEmail controller
The getContactsByEmail controller uses the keyword "findOne". So, the controller only gets one contact for the email. But the ContactsTable() page in the CRM-frontend should display all the contacts for the email. So I changed it to "find" to get all the contacts for the email.

Also, I changed the word "contact" to +contacts".