Update contactController.js and contact.js model, index.js, status.js
Add a controller called updateContactStatus to contactController.js to update the status of contacts.

Set the status value in the contact.js model to be obtained from CONTACT_STATUS.

Also, add the router path for updateStatus to index.js.

Add COMPLETED and NOT_COMPLETE statuses to CONTACT_STATUS in "src\utils\status.js".